Added the Purchase Page UI and ClickEvents. Also added more functions to the EndPayPeriod Page

+++ app/src/main/java/com/example/me/mymoneymanager/Purchases.java
@@ -21,7 +21,7 @@ public class Purchases {
     private double Amount = 0.00;
 
     @ColumnInfo(name="DateOfPurchase")
-    private double DateOfPurchase = 0;
+    private long DateOfPurchase = 0;
 
 
     public String Uid(){return Uid;}
@@ -33,7 +33,7 @@ public class Purchases {
     public double Amount (){return Amount;}
     public void Amount(double val){Amount = val;}
 
-    public double DateOfPurchase() {return DateOfPurchase;}
-    public void DateOfPurchase(double val ){DateOfPurchase = val;}
+    public long DateOfPurchase() {return DateOfPurchase;}
+    public void DateOfPurchase(long val ){DateOfPurchase = val;}
 
 }
